ShareJourneysProject/requirements.txt
asgiref==3.7.2
<<<<<<< HEAD
beautifulsoup4==4.12.3
=======
>>>>>>> c6347c8742469fe044d9f4b96bbf4a5d20e7ac2b
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.0
<<<<<<< HEAD
colorama==0.4.6
cryptography==42.0.5
Django==5.0.3
django-bootstrap4==24.3
django-ckeditor==6.7.1
django-cors-headers==4.4.0
=======
cryptography==42.0.5
Django==5.0.3
django-ckeditor==6.7.1
>>>>>>> c6347c8742469fe044d9f4b96bbf4a5d20e7ac2b
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.15.0
drf-yasg==1.21.7
<<<<<<< HEAD
inflection==0.5.1
joblib==1.4.2
jwcrypto==1.5.6
multiparse==2022.1
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pycparser==2.22
Pygments==2.17.2
PyMySQL==1.1.0
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
pyvi==0.1.1
PyYAML==6.0.1
quo==2023.5.1
requests==2.31.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==69.2.0
six==1.16.0
sklearn-crfsuite==0.5.0
soupsieve==2.5
sqlparse==0.4.4
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.4
=======
idna==3.6
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pycparser==2.22
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
>>>>>>> c6347c8742469fe044d9f4b96bbf4a5d20e7ac2b
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
<<<<<<< HEAD


absl-py==2.1.0
astunparse==1.6.3
certifi==2024.6.2
charset-normalizer==3.3.2
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pillow==10.3.0
protobuf==4.25.3
Pygments==2.18.0
requests==2.32.3
rich==13.7.1
setuptools==70.1.1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-intel==2.16.2
termcolor==2.4.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
=======
>>>>>>> c6347c8742469fe044d9f4b96bbf4a5d20e7ac2b
